refactor: move Person package
