+++ src/main/java/cc/oakk/ladder/view/printer/impl/PersonPrinter.java
@@ -1,6 +1,6 @@
 package cc.oakk.ladder.view.printer.impl;
 
-import cc.oakk.ladder.model.Person;
+import cc.oakk.ladder.model.person.Person;
 import cc.oakk.ladder.view.printer.StringPrinter;
 
 public class PersonPrinter extends StringPrinter<Person> {
